Allow node transfers to happen during RUs
A node marked as ineligible can not be scheduled by the new RC in a
rolling update. For that reason, a rolling update will get stuck if one
of the nodes is marked as ineligible. Because a node transfer was not
allowed to occur during the roll, there was no way converge the update.
We will now allow node transfers to occur during rolls (by not
attempting to acquire the mutation lock).

In order to ensure we don't reduce the health of a cluster to lower than
n - 1, we cannot just check the health of the RC's pods. Some pods may
be controlled by a second RC during an RU. We will now check the health
of all pods in the cluster, including those owned by a second RC.

Rolling updates current do not count an ineligible node as healthy, so
the respective health checks and scheduling of concurrent RUs and node
transfers should not interfere or race with one another.